Rename package to pqn_node; depend on pqn-hardware
Renames src/pqnstack → src/pqn_node, rewrites all imports that used
pqnstack.{base,network,pqn} to pull from the new pqn-hardware package,
and pins pqn-hardware via a git URL in pyproject.toml.

- Distribution: pqnstack → pqn-node
- Entry point: pqn → pqn-node
- Drops hardware deps (numpy, pyfirmata2, pyzmq, thorlabs-apt-device) —
  they now come transitively via pqn-hardware.
- CLI retains app-only commands (toggle-game, daily-report); provider
  and router commands live in pqn-hardware's pqn-hw CLI.
