+++ src/pqn_node/api/routes/coordination.py
@@ -12,13 +12,13 @@ from fastapi import status
 from fastapi.responses import StreamingResponse
 from pydantic import BaseModel
 
-from pqnstack.app.api.deps import ClientDep
-from pqnstack.app.api.deps import StateDep
-from pqnstack.app.core.config import NodeRole
-from pqnstack.app.core.config import ask_user_for_follow_event
-from pqnstack.app.core.config import protocol_cancelled_event
-from pqnstack.app.core.config import settings
-from pqnstack.app.core.config import user_replied_event
+from pqn_node.api.deps import ClientDep
+from pqn_node.api.deps import StateDep
+from pqn_node.core.config import NodeRole
+from pqn_node.core.config import ask_user_for_follow_event
+from pqn_node.core.config import protocol_cancelled_event
+from pqn_node.core.config import settings
+from pqn_node.core.config import user_replied_event
 
 logger = logging.getLogger(__name__)
