+++ src/pqn_node/api/routes/qkd.py
@@ -11,19 +11,19 @@ from fastapi import HTTPException
 from fastapi import status
 from pydantic import BaseModel
 
-from pqnstack.app.api.deps import ClientDep
-from pqnstack.app.api.deps import StateDep
-from pqnstack.app.core.config import NodeRole
-from pqnstack.app.core.config import NodeState
-from pqnstack.app.core.config import protocol_cancelled_event
-from pqnstack.app.core.config import qkd_result_received_event
-from pqnstack.app.core.config import settings
-from pqnstack.constants import BasisBool
-from pqnstack.constants import QKDEncodingBasis
-from pqnstack.network.client import Client
+from pqn_node.api.deps import ClientDep
+from pqn_node.api.deps import StateDep
+from pqn_node.core.config import NodeRole
+from pqn_node.core.config import NodeState
+from pqn_node.core.config import protocol_cancelled_event
+from pqn_node.core.config import qkd_result_received_event
+from pqn_node.core.config import settings
+from pqn_node.constants import BasisBool
+from pqn_node.constants import QKDEncodingBasis
+from pqn_hardware.network.client import Client
 
 if TYPE_CHECKING:
-    from pqnstack.base.instrument import RotatorInstrument
+    from pqn_hardware.instrument import RotatorInstrument
 
 logger = logging.getLogger(__name__)
